+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Showcase
 
+* *[v0.11.1-rc1]*
+  * [FIX]: Ensure the user does not get stuck on the password reset page
+    after successfully resetting their password
+
 ## v0.11.0 (June 21, 2016)
 * *[v0.11.0-rc3]*
   * Add the industry CTR to the dashboard chart